# Performance - ClassMappings : The computed mapping is now added to the cache