- This bug fix allows a major performance improvement (as expected when
  using a cache)