@lauriharpf 2015-08-02
Hi @gdarmont, thanks for the pull request!

Do you happen to have an example test that demonstrates the performance improvement after applying this change? I ran a quick, highly simplistic and non-scientific test ([http://pastebin.com/082iSukr](test source is in PasteBin)) three times with `mvn -Dtest=MappingPerformance test` :

With pull request #254 added to codebase:

    Done mapping 1000000 objects. Duration 4714 ms.
    Done mapping 1000000 objects. Duration 2623 ms.

    Done mapping 1000000 objects. Duration 4247 ms.
    Done mapping 1000000 objects. Duration 2334 ms.

    Done mapping 1000000 objects. Duration 4415 ms.
    Done mapping 1000000 objects. Duration 2400 ms.

Without pull request #254:

    Done mapping 1000000 objects. Duration 3870 ms.
    Done mapping 1000000 objects. Duration 2049 ms.

    Done mapping 1000000 objects. Duration 3736 ms.
    Done mapping 1000000 objects. Duration 2634 ms.

    Done mapping 1000000 objects. Duration 4339 ms.
    Done mapping 1000000 objects. Duration 2417 ms.

This does not show a clear performance difference, but perhaps there are other cases where the performance differs?