@gdarmont 2017-04-03
@garethahealy Sorry I forgot that PR. No update needed here to me. The PR code is used in production on our side and works very well. 

@lauriharpf didn't see any performance update, and this is perfectly normal since the proposed patch puts in cache the result of `findInterfaceMapping`. The code used to make the benchmark does not use interface mapping.

But, as always, perfomance improvements may highly depends on use case. In our case, we have a lots of mappings (1000+), mostly generated programmatically.